add pydantic > 2 support

+++ thunder/cli/backend_cli.py
@@ -1,10 +1,12 @@
-from typing import List
+from typing import List, Union, Dict
 
 import yaml
 from rich.console import Console
 from rich.table import Table
 from typer import Abort, Argument, Option, Typer
 from typing_extensions import Annotated
+
+from ..pydantic_compat import model_validate, model_dump
 
 from ..backend import MetaEntry
 from .backend import BACKENDS_CONFIG_PATH, BackendEntryConfig, load_backend_configs
@@ -21,6 +23,9 @@ BackendParamsArg = Annotated[List[str], Argument(
 ForceAddArg = Annotated[bool, Option(
     "--force", "-f", help="Forces overwriting of the same backend in .yml file."
 )]
+CreateYMLArg = Annotated[bool, Option(
+    "--create", "-c", help="Creates .yml file with stored backends."
+)]
 
 console = Console()
 backend_app = Typer(name="backend", help="Commands for managing your backends.")
@@ -35,13 +40,14 @@ def _set(name: BackendNameArg):
                       f"available configs: {sorted(local)}`")
         raise Abort(1)
 
-    local["meta"] = MetaEntry.parse_obj({"default": name})
+    local["meta"] = model_validate(MetaEntry, {"default": name})
     with BACKENDS_CONFIG_PATH.open("w") as stream:
-        yaml.safe_dump({k: v.dict() for k, v in local.items()}, stream)
+        yaml.safe_dump({k: _dump_backend_entry(v) for k, v in local.items()}, stream)
 
 
 @backend_app.command()
-def add(name: BackendNameArg, params: BackendParamsArg, force: ForceAddArg = False):
+def add(name: BackendNameArg, params: BackendParamsArg, force: ForceAddArg = False,
+        create_yml: CreateYMLArg = False):
     """ Add run config to the list of available configs. """
     local = load_backend_configs()
     if name in local and not force:
@@ -52,9 +58,19 @@ def add(name: BackendNameArg, params: BackendParamsArg, force: ForceAddArg = Fal
     kwargs = dict(map(lambda p: p.split("="), params))
     config = {"backend": kwargs.pop("backend", "cli"), "config": kwargs}
 
-    local.update({name: BackendEntryConfig.parse_obj(config)})
+    local.update({name: model_validate(BackendEntryConfig, config)})
+
+    if not BACKENDS_CONFIG_PATH.parent.exists() and not create_yml:
+        path = str(BACKENDS_CONFIG_PATH)
+        console.print(f"Backends storage {path} does not exist, "
+                      f"you can create it by adding --create to the command.")
+        raise Abort(1)
+
+    if create_yml:
+        BACKENDS_CONFIG_PATH.parent.mkdir(parents=True, exist_ok=True)
+
     with BACKENDS_CONFIG_PATH.open("w") as stream:
-        yaml.safe_dump({k: v.dict() for k, v in local.items()}, stream)
+        yaml.safe_dump({k: _dump_backend_entry(v) for k, v in local.items()}, stream)
 
 
 @backend_app.command()
@@ -67,7 +83,7 @@ def remove(name: BackendNameArg):
 
     local.pop(name)
     with BACKENDS_CONFIG_PATH.open("w") as stream:
-        yaml.safe_dump({k: v.dict() for k, v in local.items()}, stream)
+        yaml.safe_dump({k: _dump_backend_entry(v) for k, v in local.items()}, stream)
 
 
 @backend_app.command(name="list")
@@ -86,10 +102,17 @@ def _list(names: BackendNamesArg = None):
         console.print("These names are not among your configs:", extra)
 
     for name in sorted(set(names if names else local) - extra.union({"meta"})):
-        entry = local[name].dict()
+        entry = _dump_backend_entry(local[name])
         table.add_row(*map(str, [name, entry.get("backend", None), entry.get("config", None)]))
 
     console.print(table)
 
     if "meta" in local:
         console.print(f"[italic green]Default is [/italic green]{local['meta'].default}")
+
+
+def _dump_backend_entry(backend: BackendEntryConfig) -> Dict[str, Union[str, Dict]]:
+    entry = model_dump(backend)
+    if hasattr(backend, "config"):
+        entry["config"] = model_dump(backend.config)
+    return entry
